Handle /support/gdbsupport.init for Docker
CHANGED:
If GDB is started in a Docker instance, you do not have access to the VS Code extension folder.
Hence, the /support/gdbsupport.init file cannot be loaded from GDB directly.
As a workaround, the file is stored in a variable and splitted line for line for parsing it to GDB running in Docker later.