Send raw commands in console mode
CHANGED:
If interpreter exec console executes define, GDB switches to console input mode.
As we want to parse the gdbsupport.init file, we would end up in a locked mode if parsing in the same way as for MI2.
Hence, the commands are splitted and handled differently if they contain def as these lines typically contain the commands from gdbsupport.init